+++ lib/dradis/plugins/qualys.rb
@@ -7,5 +7,7 @@ end
 
 require 'dradis/plugins/qualys/engine'
 require 'dradis/plugins/qualys/field_processor'
-require 'dradis/plugins/qualys/importer'
 require 'dradis/plugins/qualys/version'
+
+require 'dradis/plugins/qualys/vuln/importer'
+require 'dradis/plugins/qualys/was/importer'